Feature - Adding search by city

+++ Weather App/Domain/Services/WeatherService.swift
@@ -6,5 +6,5 @@
 //
 
 protocol WeatherService {
-    func fetchForecast(forLat lat: Double, lon: Double) async throws -> Forecast
+    func fetchForecast(forLat lat: Float, lon: Float) async throws -> Forecast
 }
